refactor(pricing) Remove duplicate verbose PlanCard usage in SubscriptionSettings
