+++ src/ui/Pricing/SubscriptionSettings/SubscriptionCard/FullAccessCard.svelte
@@ -1,11 +1,13 @@
-<script>
+<script lang="ts">
   import { showIntercom } from '@/analytics/intercom'
   import Card from './Card.svelte'
+
+  export let currentPlanName: string
 </script>
 
 <Card
   yellow
-  title="Use Pro plan wisely"
+  title="Use {currentPlanName} plan wisely"
   label="You have full access!"
   action="Academy"
   subaction="Contact support"
